Create PulsarDataAndAveragedTravelTime
I kept re-reading "Decoding the Message of the Pulsars: Intelligent Communication from the Galaxy" by Paul A. LaViolette and period derivatives were mentioned. Fractions of a period per a unit of time or per a period aren't how the periods increase. It's easier and neater than that. Anyways.. this is better. Hopefully it works with real data, in a real test....